refactor the uint fileds in jsmethod
1. make bytecodeArraySize, hotnessCounter, methodId, slotSize into one field methodField.
2. change the interpreter_stub to fit the hotness counter in jsmethod.
3. fix the fail problem of mem controller test.
4. Adjust the timeout value of uinittest

+++ ecmascript/interpreter/interpreter.h
@@ -44,7 +44,7 @@ static constexpr uint32_t INVALID_ARGS_NUMBER = -1;
 
 class EcmaInterpreter {
 public:
-    static const uint32_t METHOD_HOTNESS_THRESHOLD = 512;
+    static const int16_t METHOD_HOTNESS_THRESHOLD = 512;
     enum ActualNumArgsOfCall : uint8_t { CALLARG0 = 0, CALLARG1, CALLARGS2, CALLARGS3 };
 
     static inline JSTaggedValue Execute(EcmaRuntimeCallInfo *info);
